Addressed code challenge user stories
As a user, when the page loads, I can see the 'Image of the Day', a random image selected from those taken yesterday (not by either rover as only one has images yesterday).
As a user, I can navigate to a gallery displaying images taken yesterday by either rover.
As a user, I can filter the images by camera name.
As a user, I want to see dates formatted in the style Month dd, yyyy.
As a user, I want the app to be responsive.